requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

streamlit==1.36.0
python-dotenv==1.0.1

# LangChain + HuggingFace
langchain==0.2.10
langchain-community==0.2.10
langchain-huggingface==0.0.3
huggingface-hub==0.23.4
transformers==4.41.2
accelerate==0.29.3
sentence-transformers==2.6.1

# Vector DB + PDF
faiss-cpu==1.8.0
pypdf==4.2.0

# Torch CPU wheels
torch==2.2.2+cpu
torchvision==0.17.2+cpu
torchaudio==2.2.2+cpu

# Misc
numpy==1.26.4
typing-extensions>=4.7.0